+++ internal/amf/ngap/handle_ue_context_release_request.go
@@ -107,10 +107,10 @@ func HandleUEContextReleaseRequest(ctx ctxt.Context, ran *context.AmfRan, msg *n
 			ranUe.Log.Info("Ue Context in GMM-Registered")
 			if pDUSessionResourceList != nil {
 				for _, pduSessionReourceItem := range pDUSessionResourceList.List {
-					pduSessionID := int32(pduSessionReourceItem.PDUSessionID.Value)
+					pduSessionID := uint8(pduSessionReourceItem.PDUSessionID.Value)
 					smContext, ok := amfUe.SmContextFindByPDUSessionID(pduSessionID)
 					if !ok {
-						ranUe.Log.Error("SmContext not found", zap.Int32("PduSessionID", pduSessionID))
+						ranUe.Log.Error("SmContext not found", zap.Uint8("PduSessionID", pduSessionID))
 						continue
 					}
 					response, err := consumer.SendUpdateSmContextDeactivateUpCnxState(ctx, amfUe, smContext)
