+++ internal/models/n1n2_message_transfer_request_data.go
@@ -1,7 +1,7 @@
 package models
 
 type N1N2MessageTransferReqData struct {
-	PduSessionID int32
+	PduSessionID uint8
 	NgapIeType   N2SmInfoType
 	SNssai       *Snssai
 }
