+++ internal/smf/context/ngap_handler.go
@@ -60,17 +60,17 @@ func HandlePDUSessionResourceSetupUnsuccessfulTransfer(b []byte, ctx *SMContext)
 
 	switch resourceSetupUnsuccessfulTransfer.Cause.Present {
 	case ngapType.CausePresentRadioNetwork:
-		logger.SmfLog.Warn("PDU Session Resource Setup Unsuccessful by RadioNetwork", zap.String("Cause", radioNetworkCauseString(resourceSetupUnsuccessfulTransfer.Cause.RadioNetwork.Value)), zap.String("supi", ctx.Supi), zap.Int32("pduSessionID", ctx.PDUSessionID))
+		logger.SmfLog.Warn("PDU Session Resource Setup Unsuccessful by RadioNetwork", zap.String("Cause", radioNetworkCauseString(resourceSetupUnsuccessfulTransfer.Cause.RadioNetwork.Value)), zap.String("supi", ctx.Supi), zap.Uint8("pduSessionID", ctx.PDUSessionID))
 	case ngapType.CausePresentTransport:
-		logger.SmfLog.Warn("PDU Session Resource Setup Unsuccessful by Transport", zap.String("Cause", transportCauseString(resourceSetupUnsuccessfulTransfer.Cause.Transport.Value)), zap.String("supi", ctx.Supi), zap.Int32("pduSessionID", ctx.PDUSessionID))
+		logger.SmfLog.Warn("PDU Session Resource Setup Unsuccessful by Transport", zap.String("Cause", transportCauseString(resourceSetupUnsuccessfulTransfer.Cause.Transport.Value)), zap.String("supi", ctx.Supi), zap.Uint8("pduSessionID", ctx.PDUSessionID))
 	case ngapType.CausePresentNas:
-		logger.SmfLog.Warn("PDU Session Resource Setup Unsuccessful by NAS", zap.String("Cause", nasCauseString(resourceSetupUnsuccessfulTransfer.Cause.Nas.Value)), zap.String("supi", ctx.Supi), zap.Int32("pduSessionID", ctx.PDUSessionID))
+		logger.SmfLog.Warn("PDU Session Resource Setup Unsuccessful by NAS", zap.String("Cause", nasCauseString(resourceSetupUnsuccessfulTransfer.Cause.Nas.Value)), zap.String("supi", ctx.Supi), zap.Uint8("pduSessionID", ctx.PDUSessionID))
 	case ngapType.CausePresentProtocol:
-		logger.SmfLog.Warn("PDU Session Resource Setup Unsuccessful by Protocol", zap.String("Cause", protocolCauseString(resourceSetupUnsuccessfulTransfer.Cause.Protocol.Value)), zap.String("supi", ctx.Supi), zap.Int32("pduSessionID", ctx.PDUSessionID))
+		logger.SmfLog.Warn("PDU Session Resource Setup Unsuccessful by Protocol", zap.String("Cause", protocolCauseString(resourceSetupUnsuccessfulTransfer.Cause.Protocol.Value)), zap.String("supi", ctx.Supi), zap.Uint8("pduSessionID", ctx.PDUSessionID))
 	case ngapType.CausePresentMisc:
-		logger.SmfLog.Warn("PDU Session Resource Setup Unsuccessful by Misc", zap.String("Cause", miscCauseString(resourceSetupUnsuccessfulTransfer.Cause.Misc.Value)), zap.String("supi", ctx.Supi), zap.Int32("pduSessionID", ctx.PDUSessionID))
+		logger.SmfLog.Warn("PDU Session Resource Setup Unsuccessful by Misc", zap.String("Cause", miscCauseString(resourceSetupUnsuccessfulTransfer.Cause.Misc.Value)), zap.String("supi", ctx.Supi), zap.Uint8("pduSessionID", ctx.PDUSessionID))
 	case ngapType.CausePresentChoiceExtensions:
-		logger.SmfLog.Warn("PDU Session Resource Setup Unsuccessful by ChoiceExtensions", zap.Any("Cause", resourceSetupUnsuccessfulTransfer.Cause.ChoiceExtensions), zap.String("supi", ctx.Supi), zap.Int32("pduSessionID", ctx.PDUSessionID))
+		logger.SmfLog.Warn("PDU Session Resource Setup Unsuccessful by ChoiceExtensions", zap.Any("Cause", resourceSetupUnsuccessfulTransfer.Cause.ChoiceExtensions), zap.String("supi", ctx.Supi), zap.Uint8("pduSessionID", ctx.PDUSessionID))
 	}
 
 	return nil
